refactor(docs): Reorganize code
